Provide ombt as an installable package
