+++ setup.cfg
@@ -1,0 +1,14 @@
+[extras]
+# package dependencies for optional (non-rabbitmq) messaging drivers.
+# projects can test-depend on ombt[<drivers>]
+# e.g.: ombt[kafka,amqp1]
+amqp =
+    pyngus>=2.2.0 # Apache-2.0
+kafka =
+     # Apache-2.0
+     # Apache-2.0
+
+[entry_points]
+console_scripts =
+    ombt = ombt._main:main
+    ombt2 = ombt._main:main
