+++ tox.ini
@@ -1,0 +1,16 @@
+[tox]
+envlist = py27,py36
+minversion = 2.0
+skipsdist = True
+skip_missing_interpreters = True
+skip_install = True
+
+[testenv]
+commands = flake8 --count {toxinidir}/ombt --show-source --ignore H404,H401,H405,E731
+deps = hacking>=1.1.0
+
+[testenv:py27]
+basepython = python2.7
+
+[testenv:py36]
+basepython = python3.6
